Use -1 instead of 0xffffffff. Use POP3_ERROR macro.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2845 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/mime/mimepart.cpp
@@ -750,7 +750,7 @@ const CHAR* qs::Part::getBody() const
 bool qs::Part::setBody(const CHAR* pszBody,
 					   size_t nLen)
 {
-	if (nLen == static_cast<size_t>(-1))
+	if (nLen == -1)
 		nLen = strlen(pszBody);
 	
 	xstring_ptr strBody(allocXString(pszBody, nLen));
@@ -1077,7 +1077,7 @@ bool qs::Part::create(const Part* pParent,
 	assert(pszContent);
 	assert(pnMaxPartCount);
 	
-	if (nLen == static_cast<size_t>(-1))
+	if (nLen == -1)
 		nLen = strlen(pszContent);
 	
 	clear();
